authentication/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Muhammad Waheed <115403188+Waheedsadique@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["crytography"], version = "^3.3.0"}
psycopg2 = "^2.9.9"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlmodel = "^0.0.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
